+++ src/main/java/agh/darwinworld/model/SimulationStepListener.java
@@ -28,7 +28,7 @@ public interface SimulationStepListener {
      */
     void updateAnimal(Vector2D position, int animalCount, int maxAnimalCount);
     void updateStatistics(int step, int animalCount, int plantCount, int emptyFieldCount, String popularGenotype,
-                          int averageLifetime, int averageDescendantsAmount);
+                          float averageLifetime, float averageDescendantsAmount);
 
     /**
      * Called whenever a fire is updated in the simulation.
